[Autofill] checked_cast selection indices
WebFormControlElement::SetSelection(),
WebFormControlElement::SelectionStart(),
WebFormControlElement::SelectionEnd() take/return `unsigned`, not
`size_t`.

This CL applies checked_cast<unsigned> to SetSelection() arguments and
SetSelectionStart(), SelectionEnd() return values.

Bug: 1007974
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4887473
Commit-Queue: danakj <danakj@chromium.org>
Commit-Queue: Christoph Schwering <schwering@google.com>
Auto-Submit: Christoph Schwering <schwering@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1209760}

+++ chrome/renderer/autofill/password_autofill_agent_browsertest.cc
@@ -8,6 +8,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #include "base/command_line.h"
 #include "base/containers/cxx20_erase.h"
 #include "base/functional/bind.h"
+#include "base/numerics/safe_conversions.h"
 #include "base/run_loop.h"
 #include "base/strings/strcat.h"
 #include "base/strings/string_number_conversions.h"
@@ -699,7 +700,7 @@ class PasswordAutofillAgentTest : public ChromeRenderViewTest {
     element->SetSelectionRange(value.size(), value.size());
   }
 
-  void CheckUsernameSelection(size_t start, size_t end) {
+  void CheckUsernameSelection(unsigned start, unsigned end) {
     EXPECT_EQ(start, username_element_.SelectionStart());
     EXPECT_EQ(end, username_element_.SelectionEnd());
   }
